separate rag container in local dev
