+++ env.example
@@ -1,0 +1,57 @@
+# Copy this file to env.local before running scripts/start_local_docker.sh.
+
+# Frappe / site
+SITE_NAME=tap_lms.localhost
+FRAPPE_BRANCH=version-16
+ADMIN_PASSWORD=admin
+WEB_PORT=8000
+SOCKETIO_PORT=9000
+BUSINESS_THEME_REPO=https://github.com/Midocean-Technologies/business_theme_v14.git
+
+# rag_service site
+# Mirrors dev/prod, where rag_service runs as its own Frappe site with its
+# own database, separate from tap_lms. They talk over HTTP (RAG Settings.base_url)
+# and RabbitMQ, never by sharing a DB connection.
+RAG_SITE_NAME=rag.localhost
+RAG_POSTGRES_DB=rag_lms
+
+# Postgres
+# Single Postgres *instance* (container) hosts both databases above:
+#   POSTGRES_DB     -> tap_lms site's database
+#   RAG_POSTGRES_DB -> rag_service site's database
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=postgres
+POSTGRES_DB=postgres
+POSTGRES_PORT=5432
+
+# Redis host ports. Container-internal ports stay 6379.
+REDIS_CACHE_PORT=6379
+REDIS_QUEUE_PORT=6380
+
+# CloudAMQP / RabbitMQ.
+# In CloudAMQP, the AMQP URL usually has this shape:
+# amqps://USERNAME:PASSWORD@HOST/VIRTUAL_HOST
+# Split that URL into the fields below.
+RABBITMQ_HOST=
+RABBITMQ_PORT=5671
+RABBITMQ_VIRTUAL_HOST=
+RABBITMQ_USERNAME=
+RABBITMQ_PASSWORD=
+RABBITMQ_SUBMISSION_QUEUE=
+RABBITMQ_PLAGIARISM_RESULTS_QUEUE=
+RABBITMQ_FEEDBACK_RESULTS_QUEUE=
+
+# Optional integrations. Keep these disabled for a basic local site.
+GCS_ENABLED=0
+GCS_BUCKET_NAME=
+GCS_PROJECT_ID=
+GCS_CREDENTIALS_JSON={}
+ELEVENLABS_ENABLED=0
+ELEVENLABS_API_KEY=
+VOICE_AGENT_ENABLED=0
+VOICE_AGENT_SERVICE_URL=
+VOICE_AGENT_CLIENT_ID=
+VOICE_AGENT_CLIENT_SECRET=
+VOICE_AGENT_DEFAULT_CONTACT_GROUP_ID=
+VOICE_AGENT_AGENT_ID=
+VOICE_AGENT_AUTH_TOKEN_CACHE_TTL=3600
